requirements.txt
streamlit>=1.33
openai>=1.30
python-dotenv>=1.0
pypdf>=4.2
numpy>=1.26
faiss-cpu>=1.8
sentence-transformers>=3.0
